requirements.txt
# Core ML libraries with compatible versions
tensorflow==2.13.0
numpy>=1.24.0
pandas>=2.0.3

# OCR-related libraries
paddleocr>=2.7.0.3
paddlepaddle>=2.5.2
protobuf>=3.20.0,<4.0.0
albumentations>=1.0.0

# Image processing
Pillow>=10.1.0
opencv-python>=4.8.0

# Web and API
flask>=2.3.2
requests>=2.31.0
huggingface-hub>=0.24.0

# Utilities
tqdm>=4.66.1
setuptools>=61
toolz>=0.12.0
paramiko>=3.3.1
pdf2docx>=0.5.6
safetensors>=0.4.0
# scalene==1.5.31.1
datasets>=2.16.1